# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Medalyze API
  version: 1.0.0
  description: Contrato candidato de ARQ-03. La implementación de endpoints pertenece a BE-03 a BE-10.
servers:
- url: /api/v1
  description: Base relativa
...
paths:
...
  /perfil:
    get:
      operationId: obtenerPerfil
      summary: Consultar perfil
      tags:
      - Cuenta
      responses:
        '200':
          description: Operación exitosa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Perfil'
              example:
                identificador: 11111111-1111-4111-8111-111111111111
                nombre_para_mostrar: ejemplo
                especialidades:
                - 11111111-1111-4111-8111-111111111111
                revision: 1
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error401'
              example:
                detalle:
                  codigo: CREDENCIALES_INVALIDAS
        '403':
          description: Sin permiso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error403'
              example:
                detalle:
                  codigo: SIN_PERMISO
        '404':
          description: No encontrado o recurso ajeno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404'
              example:
                detalle:
                  codigo: NO_ENCONTRADO
        '422':
          description: Validación o cálculo bloqueado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error422'
              example:
                detalle:
                  codigo: ERROR_VALIDACION
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500'
              example:
                detalle:
                  codigo: ERROR_INTERNO
      security:
      - BearerJWT: []
    patch:
      operationId: actualizarPerfil
      summary: Actualizar perfil
      tags:
      - Cuenta
      responses:
        '200':
          description: Operación exitosa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Perfil'
              example:
                identificador: 11111111-1111-4111-8111-111111111111
                nombre_para_mostrar: ejemplo
                especialidades:
                - 11111111-1111-4111-8111-111111111111
                revision: 1
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error401'
              example:
                detalle:
                  codigo: CREDENCIALES_INVALIDAS
        '403':
          description: Sin permiso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error403'
              example:
                detalle:
                  codigo: SIN_PERMISO
        '409':
          description: Conflicto
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error409'
              example:
                detalle:
                  codigo: CORREO_YA_REGISTRADO
        '422':
          description: Validación o cálculo bloqueado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error422'
              example:
                detalle:
                  codigo: ERROR_VALIDACION
        '428':
          description: Revisión requerida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error428'
              example:
                detalle:
                  codigo: REVISION_REQUERIDA
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500'
              example:
                detalle:
                  codigo: ERROR_INTERNO
      security:
      - BearerJWT: []
      parameters:
      - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PerfilActualizar'
            example:
              nombre_para_mostrar: ejemplo
              especialidades:
              - 11111111-1111-4111-8111-111111111111
  /organizacion:
    get:
      operationId: obtenerOrganizacion
      summary: Consultar organización
      tags:
      - Cuenta
      responses:
        '200':
          description: Operación exitosa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organizacion'
              example:
                identificador: 11111111-1111-4111-8111-111111111111
                nombre_comercial: ejemplo
                codigo_moneda: ejemplo
                multiplo_redondeo: '70.83'
                revision: 1
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error401'
              example:
                detalle:
                  codigo: CREDENCIALES_INVALIDAS
        '403':
          description: Sin permiso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error403'
              example:
                detalle:
                  codigo: SIN_PERMISO
        '404':
          description: No encontrado o recurso ajeno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404'
              example:
                detalle:
                  codigo: NO_ENCONTRADO
        '422':
          description: Validación o cálculo bloqueado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error422'
              example:
                detalle:
                  codigo: ERROR_VALIDACION
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500'
              example:
                detalle:
                  codigo: ERROR_INTERNO
      security:
      - BearerJWT: []
    patch:
      operationId: actualizarOrganizacion
      summary: Actualizar organización
      tags:
      - Cuenta
      responses:
        '200':
          description: Operación exitosa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organizacion'
              example:
                identificador: 11111111-1111-4111-8111-111111111111
                nombre_comercial: ejemplo
                codigo_moneda: ejemplo
                multiplo_redondeo: '70.83'
                revision: 1
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error401'
              example:
                detalle:
                  codigo: CREDENCIALES_INVALIDAS
        '403':
          description: Sin permiso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error403'
              example:
                detalle:
                  codigo: SIN_PERMISO
        '409':
          description: Conflicto
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error409'
              example:
                detalle:
                  codigo: CORREO_YA_REGISTRADO
        '422':
          description: Validación o cálculo bloqueado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error422'
              example:
                detalle:
                  codigo: ERROR_VALIDACION
        '428':
          description: Revisión requerida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error428'
              example:
                detalle:
                  codigo: REVISION_REQUERIDA
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500'
              example:
                detalle:
                  codigo: ERROR_INTERNO
      security:
      - BearerJWT: []
      parameters:
      - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizacionActualizar'
            example:
              nombre_comercial: ejemplo
              codigo_moneda: MXN
              multiplo_redondeo: ejemplo
              espaciado_por_defecto: 1
  /especialidades:
    get:
      operationId: listarEspecialidades
      summary: Listar especialidades
      tags:
      - Cuenta
      responses:
        '200':
          description: Operación exitosa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListaEspecialidades'
              example:
                elementos:
                - identificador: 11111111-1111-4111-8111-111111111111
                  clave: ejemplo
                  nombre: ejemplo
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error401'
              example:
                detalle:
                  codigo: CREDENCIALES_INVALIDAS
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500'
              example:
                detalle:
                  codigo: ERROR_INTERNO
      security:
      - BearerJWT: []
  /onboarding:
    get:
      operationId: obtenerOnboarding
      summary: Consultar progreso de onboarding
      tags:
      - Onboarding
      responses:
        '200':
          description: Operación exitosa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Onboarding'
              example:
                seleccion_parcial: {}
                completado: true
                revision: 1
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error401'
              example:
                detalle:
                  codigo: CREDENCIALES_INVALIDAS
        '403':
          description: Sin permiso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error403'
              example:
                detalle:
                  codigo: SIN_PERMISO
        '404':
          description: No encontrado o recurso ajeno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404'
              example:
                detalle:
                  codigo: NO_ENCONTRADO
        '422':
          description: Validación o cálculo bloqueado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error422'
              example:
                detalle:
                  codigo: ERROR_VALIDACION
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500'
              example:
                detalle:
                  codigo: ERROR_INTERNO
      security:
      - BearerJWT: []
    patch:
      operationId: actualizarOnboarding
      summary: Guardar progreso de onboarding
      tags:
      - Onboarding
      responses:
        '200':
          description: Operación exitosa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Onboarding'
              example:
                seleccion_parcial: {}
                completado: true
                revision: 1
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error401'
              example:
                detalle:
                  codigo: CREDENCIALES_INVALIDAS
        '409':
          description: Conflicto
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error409'
              example:
                detalle:
                  codigo: CORREO_YA_REGISTRADO
        '422':
          description: Validación o cálculo bloqueado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error422'
              example:
                detalle:
                  codigo: ERROR_VALIDACION
        '428':
          description: Revisión requerida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error428'
              example:
                detalle:
                  codigo: REVISION_REQUERIDA
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500'
              example:
                detalle:
                  codigo: ERROR_INTERNO
      security:
      - BearerJWT: []
      parameters:
      - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OnboardingActualizar'
            example:
              ruta: ASISTENTE_GUIADO
              situacion: YA_ATIENDE_EN_CONSULTA_PRIVADA
              ultimo_paso: ejemplo
              seleccion_parcial: {}
              completado: true
...
components:
  securitySchemes:
    BearerJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT de acceso con vigencia de 15 minutos.
  parameters:
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema:
        type: string
        pattern: ^\"?[0-9]+\"?$
      description: Revisión esperada del recurso.
...
  schemas:
    CampoError:
      type: object
      additionalProperties: false
      properties:
        campo:
          type: string
        mensaje:
          type: string
      required:
      - campo
      - mensaje
    BloqueoCalculo:
      type: object
      additionalProperties: false
      properties:
        codigo:
          type: string
          enum:
          - GASTO_NEGATIVO
          - EQUIPO_INCOMPLETO
          - SIN_COSTOS_INDIRECTOS_REGISTRADOS
          - SIN_PERFIL_CAPACIDAD
          - SIN_TIEMPO_ATENCION
          - DURACION_INVALIDA
          - ESPACIADO_FUERA_RANGO
          - SIN_METODO_MATERIALES
          - SIN_MESES_VALIDOS
          - INSUMO_ESPECIFICO_INVALIDO
          - RECETA_INCOMPLETA
          - IMPORTE_MATERIALES_FALTANTE
          - MATERIALES_EN_CERO
          - AJUSTE_FUERA_RANGO
        paso:
          type: integer
          minimum: 1
          maximum: 7
        mensaje:
          type: string
        identificador_recurso:
          oneOf:
          - type: string
            format: uuid
          - type: 'null'
        campo:
          oneOf:
          - type: string
          - type: 'null'
      required:
      - codigo
      - paso
      - mensaje
    DetalleError:
      type: object
      additionalProperties: false
      properties:
        codigo:
          type: string
        mensaje:
          type: string
        campos:
          type: array
          items:
            $ref: '#/components/schemas/CampoError'
          default: []
        bloqueos:
          type: array
          items:
            $ref: '#/components/schemas/BloqueoCalculo'
          default: []
        revision_actual:
          oneOf:
          - type: integer
            minimum: 0
          - type: 'null'
        request_id:
          oneOf:
          - type: string
          - type: 'null'
      required:
      - codigo
      - mensaje
...
    Error401:
      type: object
      additionalProperties: false
      properties:
        detalle:
          allOf:
          - $ref: '#/components/schemas/DetalleError'
          properties:
            codigo:
              type: string
              enum:
              - CREDENCIALES_INVALIDAS
              - SIN_AUTENTICACION
              - TOKEN_VENCIDO
      required:
      - detalle
    Error403:
      type: object
      additionalProperties: false
      properties:
        detalle:
          allOf:
          - $ref: '#/components/schemas/DetalleError'
          properties:
            codigo:
              type: string
              enum:
              - SIN_PERMISO
      required:
      - detalle
    Error404:
      type: object
      additionalProperties: false
      properties:
        detalle:
          allOf:
          - $ref: '#/components/schemas/DetalleError'
          properties:
            codigo:
              type: string
              enum:
              - NO_ENCONTRADO
      required:
      - detalle
    Error409:
      type: object
      additionalProperties: false
      properties:
        detalle:
          allOf:
          - $ref: '#/components/schemas/DetalleError'
          properties:
            codigo:
              type: string
              enum:
              - CORREO_YA_REGISTRADO
              - INSUMO_EN_USO
              - INSUMO_DUPLICADO
              - PERIODO_YA_REGISTRADO
              - CONFLICTO_REVISION
      required:
      - detalle
    Error422:
      type: object
      additionalProperties: false
      properties:
        detalle:
          allOf:
          - $ref: '#/components/schemas/DetalleError'
          properties:
            codigo:
              type: string
              enum:
              - ERROR_VALIDACION
              - CALCULO_BLOQUEADO
      required:
      - detalle
    Error428:
      type: object
      additionalProperties: false
      properties:
        detalle:
          allOf:
          - $ref: '#/components/schemas/DetalleError'
          properties:
            codigo:
              type: string
              enum:
              - REVISION_REQUERIDA
      required:
      - detalle
    Error500:
      type: object
      additionalProperties: false
      properties:
        detalle:
          allOf:
          - $ref: '#/components/schemas/DetalleError'
          properties:
            codigo:
              type: string
              enum:
              - ERROR_INTERNO
      required:
      - detalle
...
    Organizacion:
      type: object
      additionalProperties: false
      properties:
        identificador:
          type: string
          format: uuid
        nombre_comercial:
          type: string
        codigo_moneda:
          type: string
          pattern: ^[A-Z]{3}$
          default: MXN
        multiplo_redondeo:
          type: string
          pattern: ^-?\d+(\.\d{1,10})?$
          examples:
          - '70.83'
        espaciado_por_defecto:
          oneOf:
          - type: integer
            minimum: 0
            maximum: 15
          - type: 'null'
        revision:
          type: integer
          minimum: 0
      required:
      - identificador
      - nombre_comercial
      - codigo_moneda
      - multiplo_redondeo
      - revision
...
    Perfil:
      type: object
      additionalProperties: false
      properties:
        identificador:
          type: string
          format: uuid
        nombre_para_mostrar:
          type: string
        especialidades:
          type: array
          items:
            type: string
            format: uuid
        revision:
          type: integer
          minimum: 0
        correo:
          type: string
          format: email
          readOnly: true
      required:
      - identificador
      - nombre_para_mostrar
      - especialidades
      - revision
    PerfilActualizar:
      type: object
      additionalProperties: false
      properties:
        nombre_para_mostrar:
          type: string
        especialidades:
          type: array
          items:
            type: string
            format: uuid
    OrganizacionActualizar:
      type: object
      additionalProperties: false
      properties:
        nombre_comercial:
          type: string
        codigo_moneda:
          type: string
          const: MXN
        multiplo_redondeo:
          type: string
          pattern: ^\d+(\.\d{1,2})?$
        espaciado_por_defecto:
          oneOf:
          - type: integer
            minimum: 0
            maximum: 15
          - type: 'null'
    Especialidad:
      type: object
      additionalProperties: false
      properties:
        identificador:
          type: string
          format: uuid
        clave:
          type: string
        nombre:
          type: string
      required:
      - identificador
      - clave
      - nombre
    ListaEspecialidades:
      type: object
      additionalProperties: false
      properties:
        elementos:
          type: array
          items:
            $ref: '#/components/schemas/Especialidad'
      required:
      - elementos
    Onboarding:
      type: object
      additionalProperties: false
      properties:
        ruta:
          oneOf:
          - type: string
            enum:
            - ASISTENTE_GUIADO
            - CAPTURA_DIRECTA_DATOS
          - type: 'null'
        situacion:
          oneOf:
          - type: string
            enum:
            - YA_ATIENDE_EN_CONSULTA_PRIVADA
            - ESTA_POR_INICIAR
          - type: 'null'
        ultimo_paso:
          oneOf:
          - type: string
          - type: 'null'
        seleccion_parcial:
          type: object
          additionalProperties: true
        completado:
          type: boolean
        revision:
          type: integer
          minimum: 0
      required:
      - seleccion_parcial
      - completado
      - revision
    OnboardingActualizar:
      type: object
      additionalProperties: false
      properties:
        ruta:
          type: string
          enum:
          - ASISTENTE_GUIADO
          - CAPTURA_DIRECTA_DATOS
        situacion:
          type: string
          enum:
          - YA_ATIENDE_EN_CONSULTA_PRIVADA
          - ESTA_POR_INICIAR
        ultimo_paso:
          type: string
        seleccion_parcial:
          type: object
          additionalProperties: true
        completado:
          type: boolean
...
security:
- BearerJWT: []
...